fixed location of method
